# Conflicts:
#	src/cmd/compile/internal/reflectdata/reflect.go
#	src/cmd/compile/internal/types/type.go
#	src/io/io.go
#	src/net/tcpsock.go
#	src/runtime/asm_amd64.s
#	src/runtime/lock_futex.go
#	src/runtime/lockrank_on.go
#	src/runtime/malloc.go
#	src/runtime/map.go
#	src/runtime/mbarrier.go
#	src/runtime/mbitmap.go
#	src/runtime/mfixalloc.go
#	src/runtime/mgc.go
#	src/runtime/mgcpacer.go
#	src/runtime/mgcsweep.go
#	src/runtime/mheap.go
#	src/runtime/panic.go
#	src/runtime/proc.go
#	src/runtime/runtime2.go
#	src/runtime/sema.go
#	src/runtime/stubs.go
#	src/runtime/traceback.go
#	src/sync/once.go
#	src/sync/poolqueue.go

+++ src/internal/bisect/bisect.go
@@ -482,7 +482,6 @@ func (m *Matcher) stack(w Writer) bool {
 		}
 	}
 	return m.ShouldEnable(h)
-
 }
 
 // Writer is the same interface as io.Writer.
@@ -728,7 +727,7 @@ func fnvString(h uint64, x string) uint64 {
 
 func fnvUint64(h uint64, x uint64) uint64 {
 	for i := 0; i < 8; i++ {
-		h ^= uint64(x & 0xFF)
+		h ^= x & 0xFF
 		x >>= 8
 		h *= prime64
 	}
